+++ lib/swhid.rb
@@ -7,6 +7,7 @@ require_relative "swhid/objects/directory"
 require_relative "swhid/objects/revision"
 require_relative "swhid/objects/release"
 require_relative "swhid/objects/snapshot"
+require_relative "swhid/from_filesystem"
 
 module Swhid
   class Error < StandardError; end